terraform/function_scripts/requirements.txt
functions-framework==3.*
google-cloud-aiplatform
google-cloud-secret-manager
